Migrate function.py changes.
